refactor: move micro_gateway_config.py to controller (#547)
