+++ src/dashboard/apigateway/.importlinter
@@ -31,11 +31,6 @@ ignore_imports =
     # 需要重构的依赖
     apigateway.biz.esb.permissions -> apigateway.apps.esb.helpers
     apigateway.biz.esb.permissions -> apigateway.apps.esb.utils
-    apigateway.controller.distributor.helm -> apigateway.core.micro_gateway_config
-    apigateway.controller.tasks.micro_gateway -> apigateway.core.micro_gateway_config
-    apigateway.controller.helm.values_generator -> apigateway.core.micro_gateway_config
-    apigateway.controller.crds.v1beta1.convertors.stage -> apigateway.core.micro_gateway_config
-    apigateway.controller.crds.config_controller -> apigateway.core.micro_gateway_config
     apigateway.controller.tasks.syncing -> apigateway.biz.release
     apigateway.controller.crds.release_data.release_data -> apigateway.biz.gateway_jwt
     apigateway.common.audit.shortcuts -> apigateway.apps.audit.signals
