+++ src/dashboard/apigateway/apigateway/controller/crds/v1beta1/convertors/stage.py
@@ -26,7 +26,7 @@ from apigateway.controller.crds.release_data.release_data import ReleaseData
 from apigateway.controller.crds.v1beta1.convertors.base import BaseConvertor, UrlInfo
 from apigateway.controller.crds.v1beta1.models.base import PluginConfig
 from apigateway.controller.crds.v1beta1.models.gateway_stage import BkGatewayStage, BkGatewayStageSpec
-from apigateway.core.micro_gateway_config import MicroGatewayHTTPInfo
+from apigateway.controller.micro_gateway_config import MicroGatewayHTTPInfo
 from apigateway.core.models import MicroGateway
